[BUGFIX] v12 drag and drop
for drag and drop (and for other Datahandler-operations) we use
<tx_container_parent>-<colPos> as colPos and split it in our
Datahandler hooks into colPos and tx_container_parent field.
Due to https://review.typo3.org/c/Packages/TYPO3.CMS/+/75887
we cannot use "-" as delimiter for <tx_container_parent>-<colPos>
in data-colpos tag attribute which is used for drag and drop, because
value is cast to int.
So wo introduce a new integer constant as delimiter.